feat: s-expression datum transformer with generic pairs
The pair data structure is strictly mapped to the pair definition in r7rs. so that multiple pair handling can reuse the generic one, such as **ParameterFormals** and pattern syntax.
Now the parser becomes a datum transformer to tramsform datum into primary expressions
The datum is a simpler ast descripted in #53
<datum> - <primitives>
	| <list of datum>
	| <vector of datum>